+++ packages/desktop-client/src/components/modals/BudgetAutomationsModal/index.ts
@@ -1,0 +1,2 @@
+export { BudgetAutomationsModal } from './BudgetAutomationsModal';
+export { migrateTemplatesToAutomations } from './migrateTemplatesToAutomations';
